+++ frontend/src/pages/admin/AdminIndexArticle.tsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect } from 'react'
+import React, { useState, useEffect, useContext } from 'react'
 import Grid from '@mui/material/Grid'
 import { AdminTemplate } from '../../components/admin/templates/AdminTemplate'
 import { ArticleTable } from '../../features/admin/article/components/ArticleTable'
@@ -7,9 +7,12 @@ import { useDeleteArticle } from '../../features/admin/article/hooks/useDeleteAr
 import { Article } from '../../features/admin/article/types/article'
 import { CustomizedSnackbar } from '../../components/admin/elements/CustomizedSnackbar'
 import { ConfirmModal } from '../../components/admin/elements/ConfirmModal'
+import { authContext } from '../../providers/AuthProvider'
+import { ADMIN } from '../../config/userRole'
 
 // 記事一覧画面
 export const AdminIndexArticle: React.FC = () => {
+  const currentUser = useContext(authContext)
   const fetchArticlesHooks = useFetchArticles()
   const deleteArticleHooks = useDeleteArticle()
   const [isOpenConfirmModal, setIsOpenConfirmModal] = useState(false)
@@ -32,8 +35,10 @@ export const AdminIndexArticle: React.FC = () => {
 
   const executeDeleteArticle = async () => {
     if (selectedArticle) {
-      await deleteArticleHooks.deleteArticle(selectedArticle.id)
-      await fetchArticlesHooks.fetchArticles(true)
+      if (currentUser?.role == ADMIN) {
+        await deleteArticleHooks.deleteArticle(selectedArticle.id)
+        await fetchArticlesHooks.fetchArticles(true)
+      }
       setIsOpenSnackbar(true)
     } else {
       console.error('selectedArticle is undefined.')
